app/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
mangum==0.17.0
transformers==4.33.1
torch==2.0.1
Pillow==10.0.0
fugashi==1.3.0
ipadic==1.0.0